fix(tests): move test_router_azure_acompletion to llm_translation testing
